Issue #377 - use the VMUSB drivers in place rather than as copies
* Import the XLM Fera suport directly from vmusb/daqconfig
* Fix minor errors in  CXLMTimestamp
